+++ benchmarks/ecosystem/mpsc/bounded.rs
@@ -30,6 +30,7 @@ use super::adapters::AsyncChannel;
 use super::adapters::Asyncband;
 use super::adapters::BoundedMpsc;
 use super::adapters::Flume;
+use super::adapters::Kanal;
 use super::adapters::Tokio;
 use super::support::BATCH_MESSAGES;
 use super::support::BOUNDED_CAPACITY;
@@ -40,7 +41,7 @@ use super::support::RepeatedBatch;
 use super::support::RepeatedTasks;
 use crate::support::bench_context;
 
-#[divan::bench(types = [Asyncband, Tokio, AsyncChannel, Flume])]
+#[divan::bench(types = [Asyncband, Tokio, AsyncChannel, Flume, Kanal])]
 fn try_round_trip<C: BoundedMpsc>(bencher: Bencher) {
     let (sender, mut receiver) = C::channel(BOUNDED_CAPACITY);
 
@@ -50,7 +51,7 @@ fn try_round_trip<C: BoundedMpsc>(bencher: Bencher) {
     });
 }
 
-#[divan::bench(types = [Asyncband, Tokio, AsyncChannel, Flume])]
+#[divan::bench(types = [Asyncband, Tokio, AsyncChannel, Flume, Kanal])]
 fn ready_round_trip<C: BoundedMpsc>(bencher: Bencher) {
     let mut context = bench_context();
     let (sender, mut receiver) = C::channel(BOUNDED_CAPACITY);
@@ -62,7 +63,7 @@ fn ready_round_trip<C: BoundedMpsc>(bencher: Bencher) {
 }
 
 #[divan::bench(
-    types = [Asyncband, Tokio, AsyncChannel, Flume],
+    types = [Asyncband, Tokio, AsyncChannel, Flume, Kanal],
     args = PRODUCER_COUNTS,
     sample_count = 20,
     sample_size = 1,
@@ -75,7 +76,7 @@ fn concurrent<C: BoundedMpsc>(bencher: Bencher, producer_count: usize) {
 }
 
 #[divan::bench(
-    types = [Asyncband, Tokio, AsyncChannel, Flume],
+    types = [Asyncband, Tokio, AsyncChannel, Flume, Kanal],
     args = PRODUCER_COUNTS,
     sample_count = 50,
     sample_size = 1,
@@ -87,14 +88,14 @@ fn sustained<C: BoundedMpsc>(bencher: Bencher, producer_count: usize) {
     bencher.bench_local(|| batch.run());
 }
 
-#[divan::bench(types = [Asyncband, Tokio, AsyncChannel, Flume])]
+#[divan::bench(types = [Asyncband, Tokio, AsyncChannel, Flume, Kanal])]
 fn clone_drop_sender<C: BoundedMpsc>(bencher: Bencher) {
     let (sender, _receiver) = C::channel(BOUNDED_CAPACITY);
     bencher.bench_local(|| drop(black_box(sender.clone())));
 }
 
 #[divan::bench(
-    types = [Asyncband, Tokio, AsyncChannel, Flume],
+    types = [Asyncband, Tokio, AsyncChannel, Flume, Kanal],
     consts = [1, 4096],
     args = PRODUCER_COUNTS,
     sample_count = 50,
@@ -111,7 +112,7 @@ fn sustained_capacity<C: BoundedMpsc, const CAPACITY: usize>(
 }
 
 #[divan::bench(
-    types = [Asyncband, Tokio, AsyncChannel, Flume],
+    types = [Asyncband, Tokio, AsyncChannel, Flume, Kanal],
     consts = [1, 64, 4096],
     args = [(1, 0), (4, 0), (1, 4), (4, 4), (8, 4)],
     sample_count = 50,
@@ -128,7 +129,7 @@ fn scheduled<C: BoundedMpsc, const CAPACITY: usize>(
 }
 
 #[divan::bench(
-    types = [Asyncband, Tokio, AsyncChannel, Flume],
+    types = [Asyncband, Tokio, AsyncChannel, Flume, Kanal],
     consts = [64, 4096],
     args = [1, 8],
     sample_count = 50,